Rename components to more accurate name now that they only contain one exported component
