+++ packages/studio-base/src/components/AppBar/AddPanelMenu.tsx
